requirements.txt
Babel==2.12.1
bcrypt==3.1.6
bootstrap4==0.1.0
boto3==1.11.5
botocore
boto==2.49.0
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
collectfast==2.2.0
cssselect2==0.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==4.0.6
django-allauth==0.42.0
django-bootstrap==0.2.4
django-bootstrap4==22.1
django-braces==1.15.0
django-cors-headers==3.13.0
django-coverage-plugin==1.8.0
django-crispy-forms==1.9.2
django-debug-toolbar==3.5.0
django-environ==0.4.5
django-extensions==3.2.0
django-model-utils==4.0.0
django-multiselectfield==0.1.12
django-phonenumber-field==6.3.0
django-preventconcurrentlogins==0.8.2
django-rest-auth==0.9.5
django-sslserver==0.22
django-storages==1.12.3
django-stubs==1.13.1
django-taggit==3.0.0
django-timezone-field==4.1.2
djangorestframework==3.13.1
docutils==0.15.2
google-api-python-client==1.12.8
html5lib==1.0.1
idna==2.8
jmespath==0.9.4
numpy==1.24.1
pendulum==2.0.5
phonenumbers==8.10.10
Pillow==9.4.0
psycopg2-binary==2.9.5
pycparser==2.19
pyOpenSSL==23.0.0
Pyphen==0.9.5
python-cards==1.3
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.1
pytzdata==2019.3
requests
regex==2023.8.8
s3transfer==0.3.1
six==1.15.0
sqlparse==0.3.0
stripe==2.35.1
timeago==1.0.10
tinycss2==1.0.2
typed-ast==1.5.4
urllib3==1.25.7
webdriver-manager==3.8.0
webencodings==0.5.1
whitenoise==5.0.1
collectfast==2.2.0
gunicorn==20.1.0
jinja2==3.0.3